backend/requirements.txt
django == 4.2.1
django-cors-headers==3.8.0
djangorestframework == 3.14.0
djangorestframework-simplejwt == 5.2.2
gunicorn == 20.1.0
django-filter == 23.2
Pillow==9.5.0
PyJWT==2.6.0
asgiref==3.6.0
django-rest-framework==0.1.0
djoser==2.2.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytest-django==4.4.0
requests==2.30.0
sqlparse==0.4.4
drf_extra_fields==3.4.1
pytest==7.3.1
